requirements.txt
# 依存関係の管理

# --- MCP通信ライブラリ ---
mcp==1.22.0

# --- API情報の管理 ---
pydantic-settings==2.12.0

# --- GitHubクローン用 ---
GitPython==3.1.41          

# --- AWS SDK ---
aiobotocore==2.13.1
boto3

# --- Docker操作用ライブラリ ---
docker==7.1.0

# --- LlamaIndex (RAG & LLM Logic) ---
llama-index-core==0.14.8   
llama-index-readers-file==0.5.5   

# --- AWS Bedrock Integrations ---
llama-index-llms-bedrock==0.4.2
llama-index-embeddings-bedrock==0.7.2